Restructure components into separate folders and subcomponents under components
