+++ src/App.js
@@ -5,14 +5,14 @@ import { connect } from 'react-redux';
 import './App.css';
 import AppReducer from './AppReducer';
 import * as actions from './AppActions';
-import Canvas from './Canvas';
-import BottomBar from './BottomBar';
-import NicknameMenu from './NicknameMenu';
-import MessageBox from './MessageBox';
-import KickDialog from './KickDialog';
-import LoadingScreen from './LoadingScreen';
-import CreditsMenu from './CreditsMenu';
-import AdminMenu from './AdminMenu'
+import Canvas from './Components/Canvas/Canvas';
+import BottomBar from './Components/BottomBar/BottomBar';
+import NicknameMenu from './Components/NicknameMenu/NicknameMenu';
+import MessageBox from './Components/MessageBox/MessageBox';
+import KickDialog from './Components/KickDialog/KickDialog';
+import LoadingScreen from './Components/LoadingScreen/LoadingScreen';
+import CreditsMenu from './Components/CreditsMenu/CreditsMenu';
+import AdminMenu from './Components/AdminMenu/AdminMenu'
 
 export const store = createStore(AppReducer);
