+++ src/Components/Canvas/Components/PixelInfo.css
