+++ src/Components/Canvas/Components/PixelInfo.js
